# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/v1/services": {
      "post": {
        "operationId": "ServicesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new service",
        "tags": [
          "Additional Services"
        ]
      },
      "get": {
        "operationId": "ServicesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string",
              "enum": [
                "RESTAURANT",
                "SPA",
                "TRANSPORT",
                "LAUNDRY",
                "ROOM_SERVICE",
                "OTHER"
              ]
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all services with pagination and filters",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/category/{category}": {
      "get": {
        "operationId": "ServicesController_getByCategory_v1",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "RESTAURANT",
                "SPA",
                "TRANSPORT",
                "LAUNDRY",
                "ROOM_SERVICE",
                "OTHER"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Services in category"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get services by category",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          },
          "404": {
            "description": "Service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get service by ID",
        "tags": [
          "Additional Services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully"
          },
          "404": {
            "description": "Service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update service",
        "tags": [
          "Additional Services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully"
          },
          "404": {
            "description": "Service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete service (soft delete)",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/reservation": {
      "post": {
        "operationId": "ServicesController_addToReservation_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddServiceToReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service added to reservation"
          },
          "404": {
            "description": "Reservation or service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add service to reservation",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/reservation/{id}": {
      "delete": {
        "operationId": "ServicesController_removeFromReservation_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Reservation service UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service removed from reservation"
          },
          "404": {
            "description": "Reservation service not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove service from reservation",
        "tags": [
          "Additional Services"
        ]
      }
    },
    "/api/v1/v1/services/reservation/{reservationId}": {
      "get": {
        "operationId": "ServicesController_getReservationServices_v1",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "description": "Reservation UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation services"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get services for a reservation",
        "tags": [
          "Additional Services"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotel Management API",
    "description": "API for hotel management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Room Service Breakfast",
            "description": "Service name"
          },
          "description": {
            "type": "string",
            "example": "Full breakfast delivered to your room"
          },
          "price": {
            "type": "number",
            "example": 25,
            "description": "Service price"
          },
          "category": {
            "type": "string",
            "enum": [
              "RESTAURANT",
              "SPA",
              "TRANSPORT",
              "LAUNDRY",
              "ROOM_SERVICE",
              "OTHER"
            ],
            "description": "Service category"
          },
          "active": {
            "type": "boolean",
            "default": true,
            "description": "Whether service is active"
          }
        },
        "required": [
          "name",
          "price",
          "category"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Room Service Breakfast",
            "description": "Service name"
          },
          "description": {
            "type": "string",
            "example": "Full breakfast delivered to your room"
          },
          "price": {
            "type": "number",
            "example": 25,
            "description": "Service price"
          },
          "category": {
            "type": "string",
            "enum": [
              "RESTAURANT",
              "SPA",
              "TRANSPORT",
              "LAUNDRY",
              "ROOM_SERVICE",
              "OTHER"
            ],
            "description": "Service category"
          },
          "active": {
            "type": "boolean",
            "default": true,
            "description": "Whether service is active"
          }
        }
      },
      "AddServiceToReservationDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "description": "Reservation ID"
          },
          "serviceId": {
            "type": "string",
            "description": "Service ID"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "default": 1,
            "description": "Quantity"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          }
        },
        "required": [
          "reservationId",
          "serviceId"
        ]
      },
...
    }
  }
}